Moves RunDeamon into sidecar.go, refactors updateJWTSVID method.
